Backend/routers/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
SpeechRecognition==3.10.4
PyPDF2==3.0.1
python-docx==1.1.2
pydub==0.25.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
